Solved

+++ 1026 To Carry or not to Carry/main.cpp
@@ -3,11 +3,11 @@ using namespace std;
 
 int main()
 {
-    int a, b, result;
-    while(scanf("%d %d", &a, &b) != EOF)
+    long long a, b, result;
+    while(scanf("%lld %lld", &a, &b) != EOF)
     {
         result = a ^ b;
-        printf("%d\n", result);
+        printf("%lld\n", result);
     }
     return 0;
 }
